fix vector selection at multigrid matrix assembly
Previous code implicitly assumed that input vector names of integrators are
in the same order as in the AnyData in object. However if a Matrix
Integrator for a nonlinear problem requests the Newton Iterate entry, this
will be _after_ the Newton Residual entry in the AnyData object generated by
the Newton algorithm. Thus, the Matrix Integrator was using the Newton
Residual object instead of the Newton Iterate.